requirements.txt
torch~=2.2
torchvision~=0.17
pyyaml~=6.0
wandb~=0.16
matplotlib~=3.9